// packages/microservices/wallets/src/orders/order.dto.ts
// Data Transfer Object

import { OrderType } from '@prisma/client'

export class InitTransactionDTO {
  asset_id: string
  wallet_id: string
  shares: number
  price: number
  type: OrderType
}
// class-transformer class-validator

export class InputExecuteTransactionDTO {
  order_id: string
  status: 'OPEN' | 'CLOSED'
  related_investor_id: string
  broker_transaction_id: string
  negotiated_shares: number
  price: number
}